+++ mpire/progress_bar.py
@@ -1,8 +1,7 @@
 import logging
-import multiprocessing as mp
-import multiprocessing.context
 import sys
 from datetime import datetime, timedelta
+from threading import Event, Thread
 from typing import Any, Dict, Optional, Type
 
 from tqdm.auto import tqdm
@@ -12,7 +11,7 @@ from mpire.dashboard.connection_utils import (DashboardConnectionDetails, get_da
                                               set_dashboard_connection)
 from mpire.insights import WorkerInsights
 from mpire.params import WorkerMapParams, WorkerPoolParams
-from mpire.signal import DisableKeyboardInterruptSignal, ignore_keyboard_interrupt
+from mpire.signal import DisableKeyboardInterruptSignal
 from mpire.tqdm_utils import TqdmConnectionDetails, TqdmManager
 from mpire.utils import format_seconds
 
@@ -37,11 +36,10 @@ DATETIME_FORMAT = "%Y-%m-%d, %H:%M:%S"
 
 class ProgressBarHandler:
 
-    def __init__(self, ctx: multiprocessing.context.BaseContext, pool_params: WorkerPoolParams,
-                 map_params: WorkerMapParams, show_progress_bar: bool, progress_bar_total: int,
-                 progress_bar_position: int, worker_comms: WorkerComms, worker_insights: WorkerInsights) -> None:
+    def __init__(self, pool_params: WorkerPoolParams, map_params: WorkerMapParams, show_progress_bar: bool,
+                 progress_bar_total: int, progress_bar_position: int, worker_comms: WorkerComms,
+                 worker_insights: WorkerInsights) -> None:
         """
-        :param ctx: Multiprocessing context
         :param pool_params: WorkerPool parameters
         :param map_params: Map parameters
         :param show_progress_bar: When ``True`` will display a progress bar
@@ -51,10 +49,6 @@ class ProgressBarHandler:
         :param worker_comms: Worker communication objects (queues, locks, events, ...)
         :param worker_insights: WorkerInsights object which stores the worker insights
         """
-        # When the threading backend is used we switch to a multiprocessing context, because the progress bar handler
-        # needs to be a process, not a thread. This is because when using threading, the progress bar updates will
-        # interfere too much with the main process.
-        self.ctx = mp if pool_params.start_method == 'threading' else ctx
         self.show_progress_bar = show_progress_bar
         self.progress_bar_total = progress_bar_total
         self.progress_bar_position = progress_bar_position
@@ -66,8 +60,8 @@ class ProgressBarHandler:
         else:
             self.function_details = None
 
-        self.process = None
-        self.process_started = self.ctx.Event()
+        self.thread = None
+        self.thread_started = Event()
         self.progress_bar_id = None
         self.dashboard_dict = None
         self.dashboard_details_dict = None
@@ -75,44 +69,40 @@ class ProgressBarHandler:
 
     def __enter__(self) -> 'ProgressBarHandler':
         """
-        Enables the use of the ``with`` statement. Starts a new progress handler process if a progress bar should be
+        Enables the use of the ``with`` statement. Starts a new progress handler thread if a progress bar should be
         shown
 
         :return: self
         """
         if self.show_progress_bar:
 
-            # Disable the interrupt signal. We let the process die gracefully
+            # Disable the interrupt signal. We let the thread die gracefully
             with DisableKeyboardInterruptSignal():
-
-                # We start a new process because updating the progress bar in a thread can slow down processing
-                # of results and can fail to show real-time updates
                 logger.debug("Starting progress bar handler")
-                self.process = self.ctx.Process(target=self._progress_bar_handler,
-                                                args=(TqdmManager.get_connection_details(),
-                                                      get_dashboard_connection_details(),))
-                self.process.start()
-                self.process_started.wait()
+                self.thread = Thread(target=self._progress_bar_handler, args=(TqdmManager.get_connection_details(),
+                                                                              get_dashboard_connection_details()))
+                self.thread.start()
+                self.thread_started.wait()
 
         return self
 
     def __exit__(self, exc_type: Type, *_) -> None:
         """
-        Enables the use of the ``with`` statement. Terminates the progress handler process if there is one
+        Enables the use of the ``with`` statement. Terminates the progress handler thread if there is one
         """
-        if self.show_progress_bar and self.process.is_alive():
+        if self.show_progress_bar and self.thread.is_alive():
 
             # If this exit is called with an exception, then we assume an external kill signal was received (this is,
             # for example, necessary in nested pools when an error occurs)
             if exc_type is not None:
-                self.worker_comms.set_kill_signal_received()
+                self.worker_comms.signal_kill_signal_received()
 
-            # Insert poison pill and close the handling process
+            # Signal shutdown and close the handling thread
             if not self.worker_comms.exception_thrown():
-                logger.debug("Adding poison pill to progress bar")
-                self.worker_comms.add_progress_bar_poison_pill()
+                logger.debug("Signalling progress bar to shut down")
+                self.worker_comms.signal_progress_bar_shutdown()
             logger.debug("Joining progress bar handler")
-            self.process.join()
+            self.thread.join()
             logger.debug("Progress bar handler joined")
 
     def _progress_bar_handler(self, tqdm_connection_details: TqdmConnectionDetails,
@@ -124,10 +114,7 @@ class ProgressBarHandler:
         :param dashboard_connection_details: Dashboard manager host, port_nr and whether a dashboard is
             started/connected
         """
-        ignore_keyboard_interrupt()  # For Windows compatibility
-
         logger.debug("Progress bar handler started")
-        self.process_started.set()
 
         # Set tqdm and dashboard connection details. This is needed for nested pools and in the case forkserver or
         # spawn is used as start method
@@ -146,20 +133,25 @@ class ProgressBarHandler:
             in_notebook = 'IPKernelApp' in sys.modules['IPython'].get_ipython().config
         except (AttributeError, KeyError):
             in_notebook = False
-        if in_notebook:
+        if in_notebook and not main_progress_bar:
             print(' ', end='', flush=True)
 
         # Create progress bar and register the start time
+        tqdm.monitor_interval = False
         progress_bar = tqdm(total=self.progress_bar_total, position=self.progress_bar_position, dynamic_ncols=True,
-                            leave=True)
+                            leave=True, mininterval=0.1, maxinterval=0.5)
         self.start_t = datetime.fromtimestamp(progress_bar.start_t)
+
+        # Notify that the main process can continue working. We set it after the progress bar has been created, instead
+        # of right after this thread has started, for a better user experience
+        self.thread_started.set()
 
         # Register progress bar to dashboard in case a dashboard is started
         self._register_progress_bar(progress_bar)
 
         while True:
             # Wait for a job to finish
-            tasks_completed, from_queue = self.worker_comms.get_tasks_completed_progress_bar()
+            tasks_completed = self.worker_comms.get_tasks_completed_progress_bar()
 
             # If we received a poison pill, we should quit right away. We do force a final refresh of the progress bar
             # to show the latest status
@@ -171,11 +163,11 @@ class ProgressBarHandler:
                 if self.worker_comms.exception_thrown() or self.worker_comms.kill_signal_received():
                     progress_bar.set_description('Exception occurred, terminating ... ')
                     if self.worker_comms.exception_thrown():
-                        _, traceback_str = self.worker_comms.get_exception()
-                        self._send_update(progress_bar, failed=True, traceback_str=traceback_str)
+                        _, traceback_str = self.worker_comms.get_exception(in_thread=True)
+                        self._send_dashboard_update(progress_bar, failed=True, traceback_str=traceback_str)
                         self.worker_comms.task_done_exception()
                     elif self.worker_comms.kill_signal_received():
-                        self._send_update(progress_bar, failed=True, traceback_str='Kill signal received')
+                        self._send_dashboard_update(progress_bar, failed=True, traceback_str='Kill signal received')
 
                 # Final update of the progress bar. When we're not in a notebook and this is the main progress bar, we
                 # add as many newlines as the highest progress bar position, such that new output is added after the
@@ -187,27 +179,28 @@ class ProgressBarHandler:
                     progress_bar.disable = True
                     if main_progress_bar:
                         progress_bar.fp.write('\n' * (tqdm_position_register.get_highest_progress_bar_position() + 1))
-                if from_queue:
-                    self.worker_comms.task_done_progress_bar()
                 break
 
+            # Check if there's an actual update
+            elif tasks_completed > 0 and tasks_completed == progress_bar.n:
+                continue
+
             # Update progress bar
-            progress_bar.update(tasks_completed)
-            self.worker_comms.task_done_progress_bar()
+            progress_bar.update(tasks_completed - progress_bar.n)
 
             # Force a refresh when we're at 100%
             if progress_bar.n == progress_bar.total:
                 if in_notebook:
                     progress_bar.close()
                 progress_bar.refresh()
-                self.worker_comms.set_progress_bar_complete()
+                self.worker_comms.signal_progress_bar_complete()
                 self.worker_comms.wait_until_progress_bar_is_complete()
-                self._send_update(progress_bar)
+                self._send_dashboard_update(progress_bar)
 
             # Send update to dashboard in case a dashboard is started, but only when tqdm updated its view as well. This
             # will make the dashboard a lot more responsive
             if progress_bar.n == progress_bar.last_print_n:
-                self._send_update(progress_bar)
+                self._send_dashboard_update(progress_bar)
 
         logger.debug("Progress bar handler done")
 
@@ -227,10 +220,11 @@ class ProgressBarHandler:
             dashboard_tqdm_lock.acquire()
             self.progress_bar_id = len(self.dashboard_dict.keys()) + 1
             self.dashboard_details_dict.update([(self.progress_bar_id, self.function_details)])
-            self._send_update(progress_bar)
+            self._send_dashboard_update(progress_bar)
             dashboard_tqdm_lock.release()
 
-    def _send_update(self, progress_bar: tqdm, failed: bool = False, traceback_str: Optional[str] = None) -> None:
+    def _send_dashboard_update(self, progress_bar: tqdm, failed: bool = False,
+                               traceback_str: Optional[str] = None) -> None:
         """
         Adds a progress bar update to the shared dict so the dashboard process can use it, only when a dashboard has
         started